changed update behaviour
better update / always clean (config auto update still needed)

+++ lethal-company/update.ps1
@@ -1,5 +1,12 @@
-﻿cd $env:ProgramData\ZeroModded
-Remove-Item -Path .\Zeros-LethalModder.exe -Force
-Move-Item -Path .\Zeros-LethalModder_New.exe -Destination .\Zeros-LethalModder.exe -Force
+﻿Set-Location "C:\ProgramData\ZeroModded"
+if(Get-Process | Where-Object {$_.ProcessName -eq "Zeros-LethalModder"}) {
+    Get-Process | Where-Object {$_.ProcessName -eq "Zeros-LethalModder"} | Stop-Process -Force
+    while(Get-Process | Where-Object {$_.ProcessName -eq "Zeros-LethalModder"}) {
+        Sleep(1)
+    }
+}
+if(Test-Path -Path .\Zeros-LethalModder.exe) {
+    Remove-Item -Path .\Zeros-LethalModder.exe -Force
+}
+Invoke-WebRequest "https://github.com/ZeroTw0016/published-programms/raw/main/lethal-company/Zeros-LethalModder.exe" -OutFile Zeros-LethalModder.exe
 .\Zeros-LethalModder.exe
-Remove-Item .\update.exe -Force
